[refactor] AxiosMock파일 위치 이동
- common폴더 없앰으로 인한 utils폴더로의 이동
